--- packages/bundle/__tests__/types/__typescript__/fail-once/invalid-prop-type-dir.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-import ReactWebChat from '../../../../lib/index';
-
-// "dir" must be a string of "ltr' | 'rtl' | 'auto'.
-<ReactWebChat dir={123} directLine={undefined} />;
