--- packages/bundle/__tests__/types/__typescript__/fail-once/invalid-style-options-suggested-actions-stacked-overflow.ts
@@ -1,4 +1,0 @@
-import { createStyleSet } from '../../../../lib/index';
-
-// Related to #4081.
-createStyleSet({ suggestedActionsStackedOverflow: 'string' });
