--- packages/bundle/__tests__/types/__typescript__/fail-once/invalid-type-for-use-style-options.tsx
@@ -1,5 +1,0 @@
-import { hooks } from '../../../../lib/index';
-
-const [styleOptions] = hooks.useStyleOptions();
-
-styleOptions.cardEmphasisBackgroundColor = 123;
