--- packages/bundle/__tests__/types/__typescript__/pass/additional-style-options-in-full.ts
@@ -1,3 +1,0 @@
-import { createStyleSet } from '../../../../lib/index';
-
-createStyleSet({ cardEmphasisBackgroundColor: 'orange' });
