--- packages/bundle/__tests__/types/__typescript__/pass/import-component-send-text-box.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import { Components } from '../../../../lib/index';
-
-const { SendTextBox } = Components;
-
-<SendTextBox />;
