--- packages/bundle/__tests__/types/__typescript__/pass/render-with-store.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import { Components, createStore } from '../../../../lib/index';
-
-const { Composer } = Components;
-
-const store = createStore({});
-
-ReactDOM.render(<Composer directLine={undefined} store={store} />, document.getElementById('app'));
