+++ packages/bundle/test-d/fail-once/create-direct-line-invalid-bot-agent.test-d.tsx
@@ -1,0 +1,9 @@
+import { expectNotAssignable } from 'tsd';
+
+import { createDirectLine } from '../../src/index';
+
+// "botAgent" is a forbidden option.
+type CreateDirectLineInit = Parameters<typeof createDirectLine>[0];
+
+// Equivalent to: createDirectLine({ botAgent: '123' });
+expectNotAssignable<CreateDirectLineInit>({ botAgent: '123' });
