+++ packages/bundle/test-d/fail-once/invalid-type-for-use-style-options.test-d.tsx
@@ -1,0 +1,8 @@
+import { expectNotAssignable } from 'tsd';
+
+import { hooks } from '../../src/index';
+
+const [styleOptions] = hooks.useStyleOptions();
+
+// eslint-disable-next-line no-magic-numbers
+expectNotAssignable<typeof styleOptions.cardEmphasisBackgroundColor>(123);
