+++ packages/bundle/test-d/fail-once/no-full-bundle-style-options-in-minimal-for-hooks.test-d.ts
@@ -1,0 +1,11 @@
+import { expectNotAssignable } from 'tsd';
+
+import { hooks } from '../../src/index-minimal';
+
+type StyleOptions = ReturnType<typeof hooks.useStyleOptions>[0];
+
+// const [styleOptions] = hooks.useStyleOptions();
+
+expectNotAssignable<StyleOptions>({ cardEmphasisBackgroundColor: 'black' });
+
+// Equivalent to: styleOptions.cardEmphasisBackgroundColor = 'black';
