+++ packages/bundle/test-d/pass/additional-style-options-in-full.test-d.ts
@@ -1,0 +1,3 @@
+import { createStyleSet } from '../../src/index';
+
+createStyleSet({ cardEmphasisBackgroundColor: 'orange' });
