+++ packages/bundle/test-d/pass/composition-with-store.test-d.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import { Components, createStore } from '../../src/index';
+
+const { BasicWebChat, Composer } = Components;
+
+const store = createStore({});
+
+ReactDOM.render(
+  <Composer directLine={undefined} store={store}>
+    <BasicWebChat />
+  </Composer>,
+  document.getElementById('app')
+);
