+++ packages/bundle/test-d/pass/correct-type-dir.test-d.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+import ReactWebChat from '../../src/index';
+
+// eslint-disable-next-line react/jsx-indent
+<ReactWebChat dir="ltr" directLine={undefined} />;
