+++ packages/bundle/test-d/pass/import-component-send-text-box.test-d.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { Components } from '../../src/index';
+
+const { SendTextBox } = Components;
+
+// eslint-disable-next-line react/jsx-indent
+<SendTextBox />;
