+++ packages/bundle/test-d/pass/import-create-direct-line.test-d.tsx
@@ -1,0 +1,3 @@
+import { createDirectLine } from '../../src/index';
+
+createDirectLine({ token: 'faketoken' });
