+++ packages/bundle/test-d/pass/render-with-minimal-bundle.test-d.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import ReactWebChat, { createDirectLine } from '../../src/index-minimal';
+
+const directLine = createDirectLine({ token: '...' });
+const styleOptions = { accent: 'black' };
+
+ReactDOM.render(<ReactWebChat directLine={directLine} styleOptions={styleOptions} />, document.getElementById('app'));
