+++ packages/bundle/test-d/pass/render-with-store.test-d.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import { Components, createStore } from '../../src/index';
+
+const { Composer } = Components;
+
+const store = createStore({});
+
+ReactDOM.render(<Composer directLine={undefined} store={store} />, document.getElementById('app'));
