+++ packages/bundle/test-d/pass/render-with-style-set.test-d.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import ReactWebChat, { createDirectLine, createStyleSet } from '../../src/index-minimal';
+
+const directLine = createDirectLine({ token: '...' });
+const styleOptions = { accent: 'black', cardEmphasisBackgroundColor: 'orange' };
+const styleSet = createStyleSet(styleOptions);
+
+ReactDOM.render(<ReactWebChat directLine={directLine} styleSet={styleSet} />, document.getElementById('app'));
