+++ packages/bundle/test-d/pass/render.test-d.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import ReactWebChat, { createDirectLine } from '../../src/index';
+
+const directLine = createDirectLine({ token: '...' });
+
+ReactDOM.render(<ReactWebChat directLine={directLine} />, document.getElementById('app'));
