+++ packages/bundle/test-d/pass/use-style-options-with-full-bundle.test-d.ts
@@ -1,0 +1,5 @@
+import { hooks } from '../../src/index';
+
+const [styleOptions] = hooks.useStyleOptions();
+
+styleOptions.cardEmphasisBackgroundColor = 'black';
