+++ packages/bundle/test-d/pass/valid-style-options-suggested-actions-stacked-overflow.test-d.ts
@@ -1,0 +1,7 @@
+import { createStyleSet } from '../../src/index';
+
+// Related to #4081.
+createStyleSet({ suggestedActionsStackedOverflow: 'auto' });
+createStyleSet({ suggestedActionsStackedOverflow: 'hidden' });
+createStyleSet({ suggestedActionsStackedOverflow: 'scroll' });
+createStyleSet({ suggestedActionsStackedOverflow: 'visible' });
