--- packages/core/__tests__/types/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true
-  }
-}
